--- Comonicon.toml
@@ -1,1 +1,0 @@
-name = "medyansimrunner"
