+++ Project.toml
@@ -4,10 +4,9 @@ authors = ["nhz2 <nhz2@cornell.edu>"]
 version = "0.3.0"
 
 [deps]
-Comonicon = "863f3e99-da2a-4334-8734-de3dacbe5542"
+ArgCheck = "dce04be8-c92d-5529-be00-80e4d2c0e197"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 DeepDiffs = "ab62b9b5-e342-54a8-a765-a90f495de1a6"
-Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
@@ -15,12 +14,10 @@ LoggingExtras = "e6f89c97-d47a-5376-807f-9c37f3926c36"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SHA = "ea8e919c-243c-51af-8825-aaa63cd721ce"
 SmallZarrGroups = "d423b6e5-1c84-4ae2-8d2d-b903aee15ac7"
-TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 
 [compat]
-Comonicon = "1"
 DeepDiffs = "1"
 JSON3 = "1"
 LoggingExtras = "1"
-SmallZarrGroups = "0.5"
+SmallZarrGroups = "0.5, 0.6"
 julia = "1.8"
