--- deps/build.jl
@@ -1,2 +1,0 @@
-using MEDYANSimRunner
-MEDYANSimRunner.comonicon_install()
